fix detecting user leaving
