+++ realscript-app/src/modules/memberList.js
@@ -1,35 +1,33 @@
 import React from "react";
 import "../assets/main.css";
-import { client } from "../modules/socketClient";
+// import LocalContext from "./localContext";
+import { client } from "./socketClient";
 
 export default class MemberList extends React.Component {
   state = {
-    // user_list: [],
-    user_activity: []
+    msgs: [],
+    userActivity: []
   };
 
   componentDidMount() {
     this._isMounted = true;
     client.onmessage = (event) => {
-      // on receiving a message, add it to the list of messages
       const data = JSON.parse(event.data).data;
       const type = JSON.parse(event.data).type;
-      console.log(data.userActivity);
-      // const usr = JSON.parse(data).name;
-      console.log("This is updating memberList!!!!");
-      console.log(type);
-      // console.log(usr);
-      // this.addUser(usr);
-      if (type === "userevent") {
-        console.log("reach the user activtiy???");
-        this.setState({ user_activity: data.userActivity });
+      if (type === "contentchange") {
+        const msg = JSON.parse(data).message;
+        console.log("content change!!!");
+        console.log(msg);
+        this.addHistory(msg);
+      } else if (type === "userevent") {
+        this.setState({ userActivity: data.userActivity });
+        console.log("user activity change!!");
+        console.log(this.state.userActivity);
       }
-      console.log(this.state.user_activity);
     };
 
     client.onclose = () => {
       console.log("Disconnected From WebSocket");
-      // this.setState({ ws: new w3cwebsocket(URL, "chatting") });
     };
   }
 
@@ -37,22 +35,20 @@ export default class MemberList extends React.Component {
     this._isMounted = false;
   }
 
-  // addUser = (usr) => {
-  //   if (!this.state.user_list.includes(usr)) {
-  //     this.setState((state) => ({ user_list: [...state.user_list, usr] }));
-  //   }
-  // };
-
   buildUsersStatus() {
     let responseCollection = "";
-    for (const usr_act of this.state.user_activity) {
+    console.log(this.state.userActivity);
+    for (const usr_act of this.state.userActivity) {
       responseCollection += usr_act + "\n";
     }
     console.log("this is responseCollection");
     return responseCollection;
   }
 
+  addHistory = (msg) => this.setState((state) => ({ msgs: [...state.msgs, msg] }));
+
   render() {
+    global.msg = this.state.msgs;
     return (
       <div className="chat-block">
         <p className="chat-title">Active Member List:</p>
@@ -63,6 +59,7 @@ export default class MemberList extends React.Component {
               key.preventDefault();
               this.props.sendMessage(this.state.messageContent);
               this.setState({ message: "" });
+              global.forceUpdate = true;
             }
           }}
         >
